PowerTool/requirements.txt
altgraph==0.17.4
canopen==2.3.0
click==8.1.7
colorama==0.4.6
et-xmlfile==1.1.0
exceptiongroup==1.2.2
importlib_metadata==8.2.0
iniconfig==2.0.0
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy==1.24.4
openpyxl==3.1.5
packaging==24.1
pandas==2.0.3
pefile==2023.2.7
pluggy==1.5.0
psutil==6.0.0
py_canoe==2.1.3
pydlt==0.3.3
pyinstaller==6.9.0
pyinstaller-hooks-contrib==2024.7
PyQt5==5.15.9
pyqt5-plugins==5.15.9.2.3
PyQt5-Qt5==5.15.2
pyqt5-tools==5.15.9.3.3
PyQt5_sip==12.15.0
pyserial==3.5
PySide2==5.15.2.1
pytest==8.3.2
pytest-html==4.1.1
pytest-metadata==3.1.1
python-can==4.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pywin32==306
pywin32-ctypes==0.2.2
qt5-applications==5.15.2.2.3
qt5-tools==5.15.2.1.3
shiboken2==5.15.2.1
six==1.16.0
tomli==2.0.1
typing_extensions==4.12.2
tzdata==2024.1
wrapt==1.16.0
zipp==3.19.2
